Refactor: DailyRecord member 도메인으로 이동
