+++ ilkda/src/main/java/com/ilkda/server/member/repository/DailyRecordRepository.java
@@ -1,6 +1,6 @@
-package com.ilkda.server.record.repository;
+package com.ilkda.server.member.repository;
 
-import com.ilkda.server.record.model.DailyRecord;
+import com.ilkda.server.member.model.DailyRecord;
 import org.springframework.data.repository.CrudRepository;
 
 public interface DailyRecordRepository extends CrudRepository<DailyRecord, Long> {
